+++ cookbooks/app-react-ag-grid/app.config.js
@@ -1,0 +1,13 @@
+// demo
+export default () => ({
+  // "appKey": "",
+  "manifest": {
+    "key": "ag-grid"
+  },
+  "environment": {},
+  "endpoints": {
+    "api": "https://foo.barz"
+  },
+  // override portal lookup
+  "portalHost": undefined
+});